wip: migrating project to clean architecture
