+++ TodoApp.Domain/Todo/TodoItem.cs
@@ -1,0 +1,9 @@
+namespace TodoApp.Domain.Todo
+{
+    public class TodoItem
+    {
+        public long Id { get; set; }
+        public string? Description { get; set; }
+        public bool State { get; set; } = false;
+    }
+}
