+++ TodoApp.sln
@@ -1,18 +1,22 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33110.190
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TodoApp", "TodoApp\TodoApp.csproj", "{E960AACA-C5E3-427B-A0DF-F060936F28A1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TodoApp", "TodoApp\TodoApp.csproj", "{E960AACA-C5E3-427B-A0DF-F060936F28A1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TodoApp.Data", "TodoApp.Data\TodoApp.Data.csproj", "{D1EE8E34-36B5-4AE2-8B57-C4547FD04CE5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TodoApp.Infrastructure", "TodoApp.Data\TodoApp.Infrastructure.csproj", "{D1EE8E34-36B5-4AE2-8B57-C4547FD04CE5}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{CDCAC7AD-CC40-4F52-98A0-E16B99AC6119}"
-ProjectSection(SolutionItems) = preProject
-	docker-compose.yml = docker-compose.yml
-	Dockerfile = Dockerfile
-	README.md = README.md
-EndProjectSection
+	ProjectSection(SolutionItems) = preProject
+		docker-compose.yml = docker-compose.yml
+		Dockerfile = Dockerfile
+		README.md = README.md
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TodoApp.Application", "TodoApp.Application\TodoApp.Application.csproj", "{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TodoApp.Domain", "TodoApp.Domain\TodoApp.Domain.csproj", "{0717771C-FE22-4BBA-B34F-096611CE972D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,9 +26,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{E960AACA-C5E3-427B-A0DF-F060936F28A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -51,5 +52,35 @@ Global
 		{D1EE8E34-36B5-4AE2-8B57-C4547FD04CE5}.Release|x64.Build.0 = Release|Any CPU
 		{D1EE8E34-36B5-4AE2-8B57-C4547FD04CE5}.Release|x86.ActiveCfg = Release|Any CPU
 		{D1EE8E34-36B5-4AE2-8B57-C4547FD04CE5}.Release|x86.Build.0 = Release|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Debug|x64.Build.0 = Debug|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Debug|x86.Build.0 = Debug|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Release|x64.ActiveCfg = Release|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Release|x64.Build.0 = Release|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Release|x86.ActiveCfg = Release|Any CPU
+		{EBE1BFC0-4B4D-491A-975D-70AD0B1EFC91}.Release|x86.Build.0 = Release|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Debug|x64.Build.0 = Debug|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Debug|x86.Build.0 = Debug|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Release|x64.ActiveCfg = Release|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Release|x64.Build.0 = Release|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Release|x86.ActiveCfg = Release|Any CPU
+		{0717771C-FE22-4BBA-B34F-096611CE972D}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EEB27A2B-39CA-47A3-A6B6-038C523FCC72}
 	EndGlobalSection
 EndGlobal
