--- TodoApp/Configuration/CorsConfiguration.cs
@@ -1,7 +1,0 @@
-namespace TodoApp.Configuration
-{
-    public class CorsConfiguration
-    {
-        public string[] AllowedOrigins { get; set; }
-    }
-}
