--- TodoApp/Configuration/ExternalConfiguration.cs
@@ -1,7 +1,0 @@
-namespace TodoApp.Configuration
-{
-    public class ExternalConfiguration
-    {
-        public string Url { get; set; }
-    }
-}
