--- TodoApp/Configuration/TodoAppConfiguration.cs
@@ -1,9 +1,0 @@
-namespace TodoApp.Configuration
-{
-    public class TodoAppConfiguration
-    {
-        public CorsConfiguration Cors { get; set; }
-
-        public ExternalConfiguration External { get; set; }
-    }
-}
