--- TodoApp/HealthCheck/DatabaseHealthCheck.cs
@@ -1,14 +1,0 @@
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Diagnostics.HealthChecks;
-
-namespace TodoApp.HealthCheck
-{
-    public class DatabaseHealthCheck: IHealthCheck
-    {
-        public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = new CancellationToken())
-        {
-            return Task.FromResult(HealthCheckResult.Healthy());
-        }
-    }
-}
