--- TodoApp/Web/Models/ErrorDetailsResponse.cs
@@ -1,10 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace TodoApp.Web.Models
-{
-    public class ErrorDetailsResponse
-    {
-        [JsonPropertyName("error")]
-        public string Error { get; set; }
-    }
-}
