+++ TodoApp/WebApi/Common/ErrorDetailsResponse.cs
@@ -1,0 +1,10 @@
+using System.Text.Json.Serialization;
+
+namespace TodoApp.WebApi.Common
+{
+    public class ErrorDetailsResponse
+    {
+        [JsonPropertyName("error")]
+        public string Error { get; set; }
+    }
+}
