+++ TodoApp/WebApi/Configuration/CorsConfiguration.cs
@@ -1,0 +1,7 @@
+namespace TodoApp.WebApi.Configuration
+{
+    public class CorsConfiguration
+    {
+        public string[] AllowedOrigins { get; set; }
+    }
+}
