+++ TodoApp/WebApi/Configuration/ExternalConfiguration.cs
@@ -1,0 +1,7 @@
+namespace TodoApp.WebApi.Configuration
+{
+    public class ExternalConfiguration
+    {
+        public string Url { get; set; }
+    }
+}
