+++ TodoApp/WebApi/Configuration/TodoAppConfiguration.cs
@@ -1,0 +1,9 @@
+namespace TodoApp.WebApi.Configuration
+{
+    public class TodoAppConfiguration
+    {
+        public CorsConfiguration Cors { get; set; }
+
+        public ExternalConfiguration External { get; set; }
+    }
+}
