+++ TodoApp/WebApi/Controllers/Todo/v1/GetSingle/TodoGetItemResponse.cs
@@ -1,0 +1,5 @@
+﻿namespace TodoApp.WebApi.Controllers.Todo.v1.GetSingle
+{
+    public class TodoGetItemResponse : TodoItemResponseBase
+    { }
+}
